refactor(dao): 외래키 설정 제거 후 테스트 최적화
