+++ src/test/java/domain/ChessTest.java
@@ -39,6 +39,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import dao.BoardDao;
 import dao.GameDao;
+import dao.fake.FakeBoardDao;
+import dao.fake.FakeGameDao;
+import domain.board.Turn;
 import domain.piece.Color;
 import domain.result.ChessResult;
 import org.junit.jupiter.api.BeforeEach;
@@ -51,11 +54,10 @@ public class ChessTest {
 
     @BeforeEach
     void setUp() {
-        BoardDao boardDao = new BoardDao();
-        boardDao.deleteAll();
-        GameDao gameDao = new GameDao();
-        gameDao.deleteAll();
-        chess = new Chess();
+        GameDao gameDao = new FakeGameDao();
+        gameDao.save(new Turn(Color.WHITE));
+        BoardDao boardDao = new FakeBoardDao();
+        chess = new Chess(gameDao, boardDao);
     }
 
     @Test
