+++ src/test/java/domain/result/ScoreCalculatorTest.java
@@ -5,6 +5,7 @@ import static domain.piece.PositionFixture.A2;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import dao.BoardDao;
+import dao.fake.FakeBoardDao;
 import domain.board.Board;
 import domain.board.Turn;
 import domain.piece.Bishop;
@@ -21,11 +22,11 @@ import org.junit.jupiter.api.Test;
 public class ScoreCalculatorTest {
 
     private ScoreCalculator scoreCalculator;
+    private BoardDao boardDao;
 
     @BeforeEach
     void setUp() {
-        BoardDao boardDao = new BoardDao();
-        boardDao.deleteAll();
+        boardDao = new FakeBoardDao();
         scoreCalculator = new ScoreCalculator();
     }
 
@@ -33,7 +34,7 @@ public class ScoreCalculatorTest {
     @DisplayName("초기 체스판의 경우 흰색 진영의 점수는 38점이다.")
     void calculate_InitWhite() {
         Turn turn = new Turn(Color.WHITE);
-        Board board = Board.create();
+        Board board = Board.create(boardDao);
 
         double score = scoreCalculator.calculate(board, turn);
 
@@ -44,7 +45,7 @@ public class ScoreCalculatorTest {
     @DisplayName("초기 체스판의 경우 검은색 진영의 점수는 38점이다.")
     void calculate_InitBlack() {
         Turn turn = new Turn(Color.BLACK);
-        Board board = Board.create();
+        Board board = Board.create(boardDao);
 
         double score = scoreCalculator.calculate(board, turn);
 
@@ -54,9 +55,8 @@ public class ScoreCalculatorTest {
     @Test
     @DisplayName("퀸은 9점으로 계산한다.")
     void calculate_Queen_Nine() {
-        Board board = Board.create(() -> {
-            BoardDao boardDao = new BoardDao();
-            boardDao.deleteAll();
+        Board board = Board.create(boardDao, () -> {
+            boardDao.delete();
             boardDao.save(A1, new Queen(Color.WHITE));
         });
 
@@ -70,9 +70,8 @@ public class ScoreCalculatorTest {
     @Test
     @DisplayName("룩은 5점으로 계산한다.")
     void calculate_Rook_Five() {
-        Board board = Board.create(() -> {
-            BoardDao boardDao = new BoardDao();
-            boardDao.deleteAll();
+        Board board = Board.create(boardDao, () -> {
+            boardDao.delete();
             boardDao.save(A1, new Rook(Color.WHITE));
         });
 
@@ -86,9 +85,8 @@ public class ScoreCalculatorTest {
     @Test
     @DisplayName("비숍은 3점으로 계산한다.")
     void calculate_Bishop_Three() {
-        Board board = Board.create(() -> {
-            BoardDao boardDao = new BoardDao();
-            boardDao.deleteAll();
+        Board board = Board.create(boardDao, () -> {
+            boardDao.delete();
             boardDao.save(A1, new Bishop(Color.WHITE));
         });
 
@@ -102,9 +100,8 @@ public class ScoreCalculatorTest {
     @Test
     @DisplayName("나이트는 2.5점으로 계산한다.")
     void calculate_Bishop_TwoPointFive() {
-        Board board = Board.create(() -> {
-            BoardDao boardDao = new BoardDao();
-            boardDao.deleteAll();
+        Board board = Board.create(boardDao, () -> {
+            boardDao.delete();
             boardDao.save(A1, new Knight(Color.WHITE));
         });
 
@@ -118,9 +115,8 @@ public class ScoreCalculatorTest {
     @Test
     @DisplayName("킹은 0점으로 계산한다.")
     void calculate_King_Zero() {
-        Board board = Board.create(() -> {
-            BoardDao boardDao = new BoardDao();
-            boardDao.deleteAll();
+        Board board = Board.create(boardDao, () -> {
+            boardDao.delete();
             boardDao.save(A1, new King(Color.WHITE));
         });
 
@@ -134,9 +130,8 @@ public class ScoreCalculatorTest {
     @Test
     @DisplayName("폰은 1점으로 계산한다.")
     void calculate_Pawn_One() {
-        Board board = Board.create(() -> {
-            BoardDao boardDao = new BoardDao();
-            boardDao.deleteAll();
+        Board board = Board.create(boardDao, () -> {
+            boardDao.delete();
             boardDao.save(A1, new Pawn(Color.WHITE));
         });
 
@@ -160,9 +155,8 @@ public class ScoreCalculatorTest {
     @Test
     @DisplayName("하나의 파일에 같은 색 폰이 2개 이상일 경우 폰의 점수는 0.5점으로 계산한다.")
     void calculate_PawnsOnSameFile_ZeroPointFive() {
-        Board board = Board.create(() -> {
-            BoardDao boardDao = new BoardDao();
-            boardDao.deleteAll();
+        Board board = Board.create(boardDao, () -> {
+            boardDao.delete();
             boardDao.save(A1, new Pawn(Color.WHITE));
             boardDao.save(A2, new Pawn(Color.WHITE));
         });
